+++ src/state/states/SensorErrorState.cpp
@@ -11,7 +11,7 @@
 #include "PidNormalState.h"
 
 void SensorErrorState::onEntry(MachineStateContext& context) {
-    context.logStateEntry(static_cast<int>(getStateId()), getStateName());
+    context.logStateEntry(getStateId(), getStateName());
     LOG(ERROR, "Sensor error detected - entering safe mode");
 
     // Disable critical operations for safety
@@ -22,7 +22,7 @@ void SensorErrorState::onEntry(MachineStateContext& context) {
 }
 
 void SensorErrorState::onExit(MachineStateContext& context) {
-    context.logStateExit(static_cast<int>(getStateId()), getStateName());
+    context.logStateExit(getStateId(), getStateName());
     LOG(INFO, "Sensor error resolved - exiting safe mode");
 
     // Re-enable normal operations
@@ -45,7 +45,7 @@ std::unique_ptr<MachineState> SensorErrorState::checkTransitions(MachineStateCon
 
     // Check emergency conditions first
     if (context.isEmergencyStop()) {
-        context.logStateTransition(static_cast<int>(getStateId()), static_cast<int>(MachineStateId::EMERGENCY_STOP), "Emergency condition during sensor error");
+        context.logStateTransition(getStateId(), MachineStateId::EMERGENCY_STOP, "Emergency condition during sensor error");
         return std::make_unique<EmergencyStopState>();
     }
 
@@ -58,11 +58,11 @@ std::unique_ptr<MachineState> SensorErrorState::checkTransitions(MachineStateCon
         if (errorDuration > RECOVERY_DELAY_MS) {
             // Check if we should return to normal PID or stay disabled
             if (context.isPidEnabled()) {
-                context.logStateTransition(static_cast<int>(getStateId()), static_cast<int>(MachineStateId::PID_NORMAL), "Sensor error resolved - returning to normal operation");
+                context.logStateTransition(getStateId(), MachineStateId::PID_NORMAL, "Sensor error resolved - returning to normal operation");
                 return std::make_unique<PidNormalState>();
             }
             else {
-                context.logStateTransition(static_cast<int>(getStateId()), static_cast<int>(MachineStateId::PID_DISABLED), "Sensor error resolved but PID disabled");
+                context.logStateTransition(getStateId(), MachineStateId::PID_DISABLED, "Sensor error resolved but PID disabled");
                 return std::make_unique<PidDisabledState>();
             }
         }
@@ -77,7 +77,7 @@ std::unique_ptr<MachineState> SensorErrorState::checkTransitions(MachineStateCon
     constexpr unsigned long MAX_ERROR_DURATION_MS = 60000; // 1 minute
 
     if (errorDuration > MAX_ERROR_DURATION_MS) {
-        context.logStateTransition(static_cast<int>(getStateId()), static_cast<int>(MachineStateId::PID_DISABLED), "Persistent sensor error - disabling PID for safety");
+        context.logStateTransition(getStateId(), MachineStateId::PID_DISABLED, "Persistent sensor error - disabling PID for safety");
         context.setPidRuntimeState(false);
         return std::make_unique<PidDisabledState>();
     }
